+++ core/test/tools/skills/skill_toolset_test.ts
@@ -1,0 +1,235 @@
+/**
+ * @license
+ * Copyright 2026 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import {
+  BaseTool,
+  BaseToolset,
+  Context,
+  InvocationContext,
+  LlmRequest,
+  ReadonlyContext,
+  Skill,
+  SkillToolset,
+} from '@google/adk';
+import {describe, expect, it, vi} from 'vitest';
+
+describe('skill_toolset', () => {
+  const mockSkill: Skill = {
+    frontmatter: {
+      name: 'test-skill',
+      description: 'A test skill',
+    },
+    instructions: 'Test instructions',
+    resources: {
+      references: {
+        'doc.md': 'Doc content',
+      },
+      assets: {
+        'image.png': Buffer.from('fake image data'),
+      },
+      scripts: {
+        'run.sh': {src: 'echo hello'},
+      },
+    },
+  };
+
+  function createMockContext(agentName = 'test-agent') {
+    return new Context({
+      invocationContext: {
+        session: {state: {}},
+        agent: {name: agentName},
+      } as unknown as InvocationContext,
+    });
+  }
+
+  describe('SkillToolset', () => {
+    it('provides default tools', async () => {
+      const toolset = new SkillToolset([mockSkill]);
+      const tools = await toolset.getTools();
+      expect(tools.map((t) => t.name)).toContain('list_skills');
+      expect(tools.map((t) => t.name)).toContain('load_skill');
+      expect(tools.map((t) => t.name)).toContain('load_skill_resource');
+    });
+
+    it('returns default tools only when no context provided', async () => {
+      const toolset = new SkillToolset([mockSkill]);
+      const tools = await toolset.getTools();
+      expect(tools.length).toBe(3);
+      expect(tools.map((t) => t.name)).toEqual([
+        'list_skills',
+        'load_skill',
+        'load_skill_resource',
+      ]);
+    });
+
+    it('returns default tools only when no skills activated', async () => {
+      const toolset = new SkillToolset([mockSkill]);
+      const context = createMockContext();
+      const tools = await toolset.getTools(context);
+      expect(tools.length).toBe(3);
+    });
+
+    it('appends instructions to LLM request', async () => {
+      const toolset = new SkillToolset([mockSkill]);
+      const llmRequest: LlmRequest = {
+        contents: [],
+        toolsDict: {},
+        liveConnectConfig: {},
+      };
+
+      await toolset.processLlmRequest(createMockContext(), llmRequest);
+
+      expect(llmRequest.config?.systemInstruction).toContain(
+        "You can use specialized 'skills'",
+      );
+      expect(llmRequest.config?.systemInstruction).toContain(
+        '<name>test-skill</name>',
+      );
+    });
+
+    it('resolves additional tools when skill is activated', async () => {
+      class DummyTool extends BaseTool {
+        constructor() {
+          super({name: 'dummy_tool', description: 'dummy'});
+        }
+        _getDeclaration() {
+          return {name: 'dummy_tool', description: 'dummy'};
+        }
+        async runAsync() {
+          return 'dummy';
+        }
+      }
+      const dummyTool = new DummyTool();
+
+      const skillWithTools: Skill = {
+        frontmatter: {
+          name: 'skill-with-tools',
+          description: 'desc',
+          metadata: {
+            adk_additional_tools: ['dummy_tool'],
+          },
+        },
+        instructions: 'instructions',
+      };
+
+      const toolset = new SkillToolset([skillWithTools], {
+        additionalTools: [dummyTool],
+      });
+
+      const mockState = {
+        get: vi.fn().mockReturnValue(['skill-with-tools']),
+      };
+
+      const context = {
+        agentName: 'test-agent',
+        state: mockState,
+      } as unknown as ReadonlyContext;
+
+      const tools = await toolset.getTools(context);
+      expect(tools.map((t) => t.name)).toContain('dummy_tool');
+    });
+
+    it('throws error when duplicate BaseTool names are provided in additionalTools', async () => {
+      class DummyTool extends BaseTool {
+        constructor(name: string) {
+          super({name, description: 'dummy'});
+        }
+        _getDeclaration() {
+          return {name: this.name, description: 'dummy'};
+        }
+        async runAsync() {
+          return 'dummy';
+        }
+      }
+      const tool1 = new DummyTool('duplicate_tool');
+      const tool2 = new DummyTool('duplicate_tool');
+
+      const skillWithTools: Skill = {
+        frontmatter: {
+          name: 'skill-with-tools',
+          description: 'desc',
+          metadata: {
+            adk_additional_tools: ['duplicate_tool'],
+          },
+        },
+        instructions: 'instructions',
+      };
+
+      const toolset = new SkillToolset([skillWithTools], {
+        additionalTools: [tool1, tool2],
+      });
+
+      const mockState = {
+        get: vi.fn().mockReturnValue(['skill-with-tools']),
+      };
+
+      const context = {
+        agentName: 'test-agent',
+        state: mockState,
+      } as unknown as ReadonlyContext;
+
+      await expect(toolset.getTools(context)).rejects.toThrow(
+        'Duplicate tool name: duplicate_tool',
+      );
+    });
+
+    it('throws error when duplicate tool names are detected via BaseToolset in additionalTools', async () => {
+      class DummyTool extends BaseTool {
+        constructor(name: string) {
+          super({name, description: 'dummy'});
+        }
+        _getDeclaration() {
+          return {name: this.name, description: 'dummy'};
+        }
+        async runAsync() {
+          return 'dummy';
+        }
+      }
+
+      class DummyToolset extends BaseToolset {
+        constructor(private mockTools: BaseTool[]) {
+          super([]);
+        }
+        override async getTools() {
+          return this.mockTools;
+        }
+        override async close() {}
+      }
+
+      const tool1 = new DummyTool('shared_name');
+      const tool2 = new DummyTool('shared_name');
+      const customToolset = new DummyToolset([tool2]);
+
+      const skillWithTools: Skill = {
+        frontmatter: {
+          name: 'skill-with-tools',
+          description: 'desc',
+          metadata: {
+            adk_additional_tools: ['shared_name'],
+          },
+        },
+        instructions: 'instructions',
+      };
+
+      const toolset = new SkillToolset([skillWithTools], {
+        additionalTools: [tool1, customToolset],
+      });
+
+      const mockState = {
+        get: vi.fn().mockReturnValue(['skill-with-tools']),
+      };
+
+      const context = {
+        agentName: 'test-agent',
+        state: mockState,
+      } as unknown as ReadonlyContext;
+
+      await expect(toolset.getTools(context)).rejects.toThrow(
+        'Duplicate tool name: shared_name',
+      );
+    });
+  });
+});
